% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\docType{data}
\name{demography}
\alias{demography}
\title{Simulated Sample Demographic Data}
\format{A data frame with 4 rows and 3 colomns}
\usage{
data(demography)
}
\description{
Demographic data for 4 random participants is provided.
}
\details{
The variables are as follows:
\itemize{
  \item \code{ID} The ID of the participant
  \item \code{Age} The age of the participant
 \item \code{Gender} The gender of the participant
}
}
\keyword{datasets}

