% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata_multiple.R
\docType{data}
\name{sampledata_multiple}
\alias{sampledata_multiple}
\title{Simulated Sample SenseWear data with 4 participants combined}
\format{A data frame with 22818 rows and 3 colomns}
\usage{
data(sampledata_multiple)
}
\description{
Simulated SenseWear physical activity data for 4 random participants including METs and Time
The variables are as follows:
\itemize{
  \item \code{Time1} The time of the recorded activity
  \item \code{METs} The  Metabolic Equivalent of Task of the recorded activity
 \item \code{ID} The ID of the participant
}
}
\keyword{datasets}

